+++ src/Components/LocaleSwitchComponent.php
@@ -9,9 +9,6 @@ use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
-/**
- * @codeCoverageIgnore
- */
 #[AsTwigComponent('locale-switch', '@EnabelLayout/components/locale-switch.html.twig')]
 class LocaleSwitchComponent
 {
@@ -24,7 +21,7 @@ class LocaleSwitchComponent
         $request = $this->requestStack->getCurrentRequest();
 
         if (!$request instanceof Request) {
-            throw new \RuntimeException('No request found.');
+            throw new \RuntimeException('No request found in RequestStack when trying to switch locale. Make sure you are calling this from a request context.');
         }
 
         $route = $request->attributes->get('_locale_switch_route', $request->attributes->get('_route'));