+++ src/Components/ThemeSwitchComponent.php
@@ -6,9 +6,6 @@ namespace Enabel\LayoutBundle\Components;
 
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
-/**
- * @codeCoverageIgnore
- */
 #[AsTwigComponent('theme-switch', '@EnabelLayout/components/theme-switch.html.twig')]
 class ThemeSwitchComponent
 {