+++ src/Controller/Favicon/BrowserconfigAction.php
@@ -8,7 +8,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-class BroswerconfigAction extends AbstractController
+class BrowserconfigAction extends AbstractController
 {
     #[Route(
         path: '/favicon/browserconfig.xml',